Some more general improvements
